requirements.txt
python-dotenv
PyYAML
pandas
numpy
pyarrow
vnstock
boto3
s3fs
